backend/requirements.txt
# bcrypt
# cryptography
# fastapi
# python-jose
# python-multipart
# SQLAlchemy
# uvicorn
# passlib
fastapi==0.115.0
   uvicorn==0.30.6
   sqlalchemy==2.0.35
   psycopg2-binary==2.9.9
   python-jose[cryptography]==3.3.0
   passlib[bcrypt]==1.7.4
   bcrypt==4.0.1
   python-dotenv==1.0.1